refactor: poses in one component
