+++ client/components/poses/stillPose.js
