+++ client/components/poses/switchPose.js
@@ -3,21 +3,21 @@ import React from 'react'
 
 const Box = posed.div({
   small: {
-    width: 50,
-    height: 150,
+    width: 30,
+    height: 200,
     backgroundColor: ({c}) => c
   },
   big: {
-    width: 50,
-    height: 200,
+    width: 30,
+    height: 50,
     backgroundColor: ({c}) => c
   }
 })
 
-class Example4 extends React.Component {
+class Example6 extends React.Component {
   state = {
-    size: 'small',
-    color: '#ededed'
+    size: 'big',
+    color: 'rgba(69, 103, 178, 1)'
   }
 
   componentDidMount() {
@@ -26,17 +26,17 @@ class Example4 extends React.Component {
         this.setState(state => ({
           size: state.size === 'small' ? 'big' : 'small'
         })),
-      590
+      1900
     )
     setInterval(
       () =>
         this.setState(state => ({
           color:
-            state.color === 'rgba(138, 185, 255, 1)'
+            state.color === 'rgba(69, 103, 178, 1)'
               ? '#ededed'
-              : 'rgba(138, 185, 255, 1)'
+              : 'rgba(69, 103, 178, 1)'
         })),
-      1550
+      900
     )
   }
 
@@ -45,4 +45,4 @@ class Example4 extends React.Component {
   }
 }
 
-export default Example4
+export default Example6
